refactor(etw): migrate etw component to v2
