+++ comp/etw/def/component.go
